requirements.txt
tensorflow>=2.15
tensorflow-datasets>=4.9
numpy>=1.26
matplotlib>=3.8
opencv-python>=4.10
pillow>=10.2
pyttsx3>=2.90

